Mitigate user losing access to wallets when app starts up and NSUserDefaults is not accessible.
If this happens on startup, the app will think the user has no wallets and ask them to create/import a wallet. It seems random and hard to reproduce. Adds a defensive check to make sure user does not have existing Realm databases for wallets and delays launch if that happens.