Fix: TokenScript files on disk, from official repos that have filenames for tokens that are not in the Realm database don't get refreshed so they will cause conflicts when a copy of the same TokenScript file is updated for another token (which is in the Realm database).
e.g. TokenScript file A is on the repo server which works for holding contracts 0xC1 and 0xC2. For some reason [1], A is written locally as 0xC1.tsml and 0xC2.tsml. Only 0xC1 is in the database. Before this commit, Only 0xC1.tsml will be refreshed. So if A changes, 0xC1.tsml and 0xC2.tsml will be different and cause a conflict. This commit changes it so that every TokenScript file is refreshed.